+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/contentshare/InAppScreenCaptureSource.swift
@@ -1,0 +1,78 @@
+//
+//  InAppScreenCaptureSource.swift
+//  AmazonChimeSDK
+//
+//  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//  SPDX-License-Identifier: Apache-2.0
+//
+
+import Foundation
+import ReplayKit
+
+/// `InAppScreenCaptureSource` is used to share  screen capture within the app. When the app is in the background,
+/// there is no sample sent to handler, and screen sharing is paused. `InAppScreenCaptureSource` is only available
+/// on iOS 11+ because of `RPScreenRecorder.startCapture(handler:completionHandler:)` method.
+/// `InAppScreenCaptureSource` does not support rotation. 
+@available(iOS 11.0, *)
+@objcMembers public class InAppScreenCaptureSource: DefaultScreenCaptureSource {
+    private var screenRecorder: RPScreenRecorder {
+        return RPScreenRecorder.shared()
+    }
+
+    public override func start() {
+        super.start()
+        if screenRecorder.isRecording {
+            stop()
+        }
+        screenRecorder.startCapture(handler: { [weak self] sampleBuffer, sampleBufferType, error in
+            guard let strongSelf = self else { return }
+            if error != nil {
+                strongSelf.logger.error(msg: "RPScreenRecorder capture error received: \(error.debugDescription)")
+            } else {
+                switch sampleBufferType {
+                case .video:
+                    strongSelf.processVideo(sampleBuffer: sampleBuffer)
+                case .audioApp, .audioMic:
+                    break
+                @unknown default:
+                    break
+                }
+            }
+        }, completionHandler: { [weak self] error in
+            guard let strongSelf = self else { return }
+            if let error = error {
+                strongSelf.logger.error(msg: "RPScreenRecorder start failed: \(error.localizedDescription)" )
+                ObserverUtils.forEach(observers: strongSelf.observers) { (observer: CaptureSourceObserver) in
+                    observer.captureDidFail(error: .systemFailure)
+                }
+            } else {
+                strongSelf.logger.info(msg: "RPScreenRecorder start succeeded.")
+                ObserverUtils.forEach(observers: strongSelf.observers) { (observer: CaptureSourceObserver) in
+                    observer.captureDidStart()
+                }
+            }
+        })
+    }
+
+    public override func stop() {
+        super.stop()
+        if !screenRecorder.isRecording {
+            logger.info(msg: "RPScreenRecorder not recording, so skipping stop")
+            return
+        }
+        screenRecorder.stopCapture { [weak self] error in
+            guard let strongSelf = self else { return }
+            if let error = error {
+                strongSelf.logger.error(msg: "RPScreenRecorder stop failed: \(error.localizedDescription)")
+                ObserverUtils.forEach(observers: strongSelf.observers) { (observer: CaptureSourceObserver) in
+                    observer.captureDidFail(error: .systemFailure)
+                }
+            } else {
+                strongSelf.logger.info(msg: "RPScreenRecorder stop succeeded.")
+                ObserverUtils.forEach(observers: strongSelf.observers) { (observer: CaptureSourceObserver) in
+                    observer.captureDidStop()
+                }
+            }
+        }
+    }
+}